# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BarberShop API",
        "description": "Appointment Scheduling System for Barbershops.",
        "contact": {
            "url": "https://github.com/matheusmrqs4"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/barber-shop/register": {
            "post": {
                "tags": [
                    "BarberShop Authentication"
                ],
                "summary": "BarberShop Register",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "BarberShop name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "BarberShop email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "BarberShop password",
                                        "type": "string"
                                    },
                                    "adress1": {
                                        "description": "BarberShop Address line 1",
                                        "type": "string"
                                    },
                                    "adress2": {
                                        "description": "BarberShop Address line 2",
                                        "type": "string"
                                    },
                                    "adress3": {
                                        "description": "BarberShop Address line 3",
                                        "type": "string"
                                    },
                                    "bio": {
                                        "description": "BarberShop Biography",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "BarberShop Phone",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/barber-shop/login": {
            "post": {
                "tags": [
                    "BarberShop Authentication"
                ],
                "summary": "BarberShop Login",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "BarberShop email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "BarberShop password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                },
                                                "token": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "500": {
                        "description": "An error occurred when trying to log in"
                    }
                }
            }
        },
        "/api/barber-shop/refresh": {
            "post": {
                "tags": [
                    "BarberShop Authentication"
                ],
                "summary": "Refresh BarberShop Token",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "Token refreshed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                },
                                                "token": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred when trying to refresh token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/barber-shop/logout": {
            "post": {
                "tags": [
                    "BarberShop Authentication"
                ],
                "summary": "Logout BarberShop",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Logout successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred when trying to logout"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}